updated config, module and payment .xml files for new type of the credit payment method, added possibility to make payment and see result, change data type of credit available from int to float

+++ app/code/AlyonaKir/Credit/Api/Data/CreditInterface.php
@@ -35,9 +35,9 @@ interface CreditInterface
 
     public function setLockCreditLimit(int $lockCreditLimit): void;
 
-    public function getCreditAvailable(): int;
+    public function getCreditAvailable(): float;
 
-    public function setCreditAvailable(int $creditAvailable): void;
+    public function setCreditAvailable(float $creditAvailable): void;
 
     public function getPurchaseStatus(): int;
